Upgraded Datacenter
Tuned nominal temperatures to get the required technology PUE value for the data provided
Added efficiency metric to estimate heat recovery from the datacenter (in MW from IT power)
Added second graph with a breakdown of the datacenter-only power consumption, among other metrics.
Now the power dial only computes the nominal IT power (not the total power) with the cooling power being computed after the fact.